optimisation de actuArticlesContextuel

+++ dmCorePlugin/plugins/sidWidgetActuPlugin/modules/actusDuCabinet/templates/_actuArticleShow.php
@@ -1,5 +1,5 @@
 <?php
-// vars : $articles, $titreBloc, $widthImage, $heightImage
+// vars : $articles, $titreBloc, $widthImage, $heightImage, $withImage
 //chargement de l'helper pour la date
 use_helper('Date');
 
@@ -9,7 +9,7 @@ echo _tag('h2.title', $titreBloc);
 if(count($articles)){
 
     foreach ($articles as $article){
-        $image = '';
+//        $image = '';
 	//affichage du contenu
 //	$articleOpts = array(
 //						'container' => 'article',
@@ -29,24 +29,23 @@ if(count($articles)){
 //	$html.= get_partial('global/schema/Thing/CreativeWork/Article', $articleOpts);
    
 
-echo _open('article', array('class' => 'itemscope Article', 'itemtype' => 'http://schema.org/Article', 'itemscope' => 'itemscope'));  
-echo _open('header', array('class' => 'contentHeader'));
-
-        if ($article->getImage()->checkFileExists() == true) {
-            echo _open('div', array('class' => 'imageFullWrapper'));
-            if($width != null) {echo  _media($article->getImage())->width($width)->set('.image itemprop="image"')->alt($article->getTitle());}
-            echo _close('div');
-        }
-echo _tag('h1', array('class' => 'title itemprop name', 'itemprop' => "name"), $article->getTitle());
-echo _tag('meta', array('content' => $titreBloc, 'itemprop' => 'articleSection'));
-echo _tag('span', array('class' => 'teaser itemprop description', 'itemprop' => 'description'), $article->getResume());
-echo _open('span.date');
-    echo __('published on').' ';
-    echo _tag('time', array('class' => 'datePublished', 'itemprop' => 'datePublished', 'pubdate' => 'pubdate', 'datetime' => format_date($article->createdAt, 'I')), format_date($article->createdAt, 'D' ));
-echo _close('span');    
-echo _close('header');
-echo _tag('section', array('class' => 'contentBody', 'itemprop' => 'articleBody'), $article->getText());
-echo _close('article');
+        echo _open('article', array('class' => 'itemscope Article', 'itemtype' => 'http://schema.org/Article', 'itemscope' => 'itemscope'));  
+            echo _open('header', array('class' => 'contentHeader'));
+                if (($withImage == true) && ($article->getImage()->checkFileExists() == true)) {
+                    echo _open('div', array('class' => 'imageFullWrapper'));
+                        if($width != null) {echo  _media($article->getImage())->width($width)->set('.image itemprop="image"')->alt($article->getTitle());}
+                    echo _close('div');
+                }
+                echo _tag('h1', array('class' => 'title itemprop name', 'itemprop' => "name"), $article->getTitle());
+                echo _tag('meta', array('content' => $titreBloc, 'itemprop' => 'articleSection'));
+                echo _tag('span', array('class' => 'teaser itemprop description', 'itemprop' => 'description'), $article->getResume());
+                echo _open('span.date');
+                    echo __('published on').' ';
+                    echo _tag('time', array('class' => 'datePublished', 'itemprop' => 'datePublished', 'pubdate' => 'pubdate', 'datetime' => format_date($article->createdAt, 'I')), format_date($article->createdAt, 'D' ));
+                echo _close('span');    
+            echo _close('header');
+            echo _tag('section', array('class' => 'contentBody', 'itemprop' => 'articleBody'), $article->getText());
+        echo _close('article');
     }
